+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/client/PaymentClient.java
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @FeignClient(name = "payment")
@@ -14,13 +15,13 @@ public interface PaymentClient {
     PaymentDto createPayment(@RequestBody OrderDto orderDto);
 
     @PostMapping("/api/v1/payment/totalCost")
-    Double totalCost(@RequestBody OrderDto orderDto);
+    BigDecimal totalCost(@RequestBody OrderDto orderDto);
 
     @PostMapping("/api/v1/payment/refund")
     void refund(@RequestBody UUID paymentId);
 
     @PostMapping("/api/v1/payment/productCost")
-    Double productCost(@RequestBody OrderDto orderDto);
+    BigDecimal productCost(@RequestBody OrderDto orderDto);
 
     @PostMapping("/api/v1/payment/failed")
     void failed(@RequestBody UUID paymentId);
