+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/warehouse/BookedProductsDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -13,8 +14,8 @@ import lombok.NoArgsConstructor;
 public class BookedProductsDto {
 
     @Positive
-    double deliveryWeight;
+    BigDecimal deliveryWeight;
     @Positive
-    double deliveryVolume;
+    BigDecimal deliveryVolume;
     boolean fragile;
 }
