+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/warehouse/DimensionDto.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -15,13 +16,13 @@ public class DimensionDto {
 
     @NotNull(message = "Необходимо указать ширину")
     @Min(value = 1, message = "минимальное значение 1")
-    Double width;
+    BigDecimal width;
 
     @NotNull(message = "Необходимо указать высоту")
     @Min(value = 1, message = "минимальное значение 1")
-    Double height;
+    BigDecimal height;
 
     @NotNull(message = "Необходимо указать глубину")
     @Min(value = 1, message = "минимальное значение 1")
-    Double depth;
+    BigDecimal depth;
 }
