+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/warehouse/NewProductWarehouseRequestDto.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -20,7 +21,7 @@ public class NewProductWarehouseRequestDto {
 
     @NotNull(message = "Вес обязателен")
     @Positive(message = "Вес должен быть положительным")
-    Double weight;
+    BigDecimal weight;
 
     @NotNull(message = "Размеры товара обязательны")
     DimensionDto dimension;
