+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -2,15 +2,16 @@ package ru.yandex.practicum.service;
 
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
 
     PaymentDto createPayment(OrderDto orderDto);
 
-    Double calculateProductCost(OrderDto orderDto);
+    BigDecimal calculateProductCost(OrderDto orderDto);
 
-    Double calculateTotalCost(OrderDto orderDto);
+    BigDecimal calculateTotalCost(OrderDto orderDto);
 
     void successPayment(UUID paymentId);
